Encapsulate menu coordinates function
