+++ frontend/src/components/card-name.jsx
@@ -1,5 +1,6 @@
 import { createSignal } from "solid-js";
 import { Menu } from "./menu";
+import { getButtonCoordinates } from "../utils";
 
 /**
  * 
@@ -18,21 +19,14 @@ export function CardName(props) {
 		props.onRenameBtnClick()
   }
 
-  function handleOptionBtnOnClick(event) {
-    event.stopPropagation();
-    const btnCoordinates = event.target.getBoundingClientRect();
-    let x = btnCoordinates.x + event.target.offsetWidth - 3;
-    const menuWidth = 82;
-    const offsetX = x + menuWidth >= window.innerWidth ? menuWidth : 0;
-    x -= offsetX;
-    const offsetY = offsetX ? 0 : 3;
-    const y = btnCoordinates.y + event.target.offsetHeight - offsetY;
-    setMenuCoordinates({ x, y });
-  }
-
   function handleMenuClose() {
     setShowMenu(false);
     setMenuCoordinates(null);
+  }
+
+  function handleOptionBtnOnClick(event) {
+    const coordinates = getButtonCoordinates(event);
+    setMenuCoordinates(coordinates);
   }
 
   const menuOptions = [
