+++ frontend/src/components/lane-name.jsx
@@ -1,5 +1,6 @@
 import { createSignal } from "solid-js";
 import { Menu } from "./menu";
+import { getButtonCoordinates } from "../utils";
 
 /**
  * 
@@ -22,15 +23,8 @@ export function LaneName(props) {
   }
 
   function handleOptionBtnOnClick(event) {
-    event.stopPropagation();
-    const btnCoordinates = event.target.getBoundingClientRect();
-    let x = btnCoordinates.x + event.target.offsetWidth - 3;
-    const menuWidth = 82;
-    const offsetX = x + menuWidth >= window.innerWidth ? menuWidth : 0;
-    x -= offsetX;
-    const offsetY = offsetX ? 0 : 3;
-    const y = btnCoordinates.y + event.target.offsetHeight - offsetY;
-    setMenuCoordinates({ x, y });
+    const coordinates = getButtonCoordinates(event);
+    setMenuCoordinates(coordinates);
   }
 
   function handleMenuClose() {
